requirements-gpu.txt
tensorflow-gpu==1.12.0
Pillow==4.2.1
requests==2.18.4
scipy==0.19.1
six==1.10.0
matplotlib==2.1.0
pytest==3.3.2
